Added pyactr
After a while of being able to run the simulations smoothly, the ImportError for pyactr suddenly popped up again. Therefore, I added it to the folder and the problem was solved. Instead it gives the same issue for simpy now, which is imported in the pyactr module... What to do about these problems...?